ai_art_education/requirements.txt
google-adk>=1.0.0
aiohttp>=3.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0